pyproject.toml
[project]
name = "sj-trading"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "polars-talib>=0.1.3",
    "polars>=1.9.0",
    "python-dotenv>=1.0.1",
    "shioaji[speed]>=1.2.5",
]

[project.scripts]
hello = "sj_trading:hello"
main = "sj_trading:main"
version = "sj_trading:show_version"
stock_testing = "sj_trading.testing_flow:testing_stock_ordering"
futures_testing = "sj_trading.testing_flow:testing_futures_ordering"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.29.5",
]
